fix: fix sparse vector sort
